single link list (part 2)

+++ c/0x01-linked_list/my_node.h
@@ -10,7 +10,7 @@
  */
 
 struct node {
-	int data;
+	long int data;
 	struct node *link;
 };
 typedef struct node my_node;
